fix: apply migration 018 and extract mention tokens to own module
- Register compound graph_edges indexes (migration 018) in Database::open —
  they were missing, causing `INDEXED BY` to error silently (swallowed by
  `unwrap_or_default`) and LinearRAG to return no results; fixes CI e2e test
- Move `extract_mention_tokens` and stopword list to `index/mentions.rs`;
  declare the module in `index/mod.rs` so it resolves correctly
